pip_requirements.txt
Flask==0.10.1
beautifulsoup4==4.3.2
click==3.3
ipython==2.3.1
requests==2.5.1
boto==2.34.0
pylint==1.4.0
flake8==2.2.5
PyYAML==3.11
